Restructure architecture
