--- src/utils/mutations.js
@@ -1,7 +1,0 @@
-import gql from 'graphql-tag'
-
-export const DELETE_NOTICE = gql`
-mutation deleteNotice($noticeId: String!) {
-  deleteNotice(noticeId: $noticeId)
-}
-`
